requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-allauth>=0.57.0
dj-rest-auth>=5.0.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
redis>=5.0.0
django-filter>=23.5
django-ckeditor>=6.7.0
drf-yasg>=1.21.7
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
better-profanity>=0.7.0
stripe>=7.0.0

